+++ src/LottoCycle.js
@@ -6,6 +6,8 @@ import Lotto from './Lotto.js';
 class LottoCycle {
   #winningNum; // 로또번호 입력값 배열, 여기가 필드
 
+  #resultWinningNum;
+
   constructor() {
     this.#winningNum = new Input();
   }
@@ -13,6 +15,11 @@ class LottoCycle {
   async run() {
     const winningNum = await this.#validCheck();
     return winningNum;
+  }
+
+  result() {
+    // bonus에 값 전달하기 위한 함수
+    return this.#resultWinningNum;
   }
 
   async #validCheck() {
@@ -32,8 +39,12 @@ class LottoCycle {
         continue;
       }
     }
-    return newWinning;
+    this.#resultWinningNum = newWinning;
+    return this.#resultWinningNum;
   }
 }
 
 export default LottoCycle;
+
+// const run = new LottoCycle();
+// run.run();
